vendor: re-vendor agent to fix mount error
The agent gRPC protocol has changed meaning the order of the gRPC
Storage message members has changed. This manifested itself with the
following error:

```
"OCI runtime create failed: rpc error: code = Internal desc = Could not resolve symlink for source 9p: unknown."
```

This was caused by the agent attempting to mount a 9p FS with a blank
mountpoint - in fact due to the change in ordering, the mountpoint was
in the mount options!

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>